chore: interface 위치 수정
